WorldCupBot/requirements.txt
discord.py>=2.3.2
asyncio
pathlib
datetime
pathlib
gspread
psutil
flask